Fix dropdown closing issues
